requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pandas>=1.5.0
numpy>=1.20.0
tqdm>=4.60.0

# LLM providers
openai>=1.0.0
anthropic>=0.7.0

# Document processing
pypdf>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
beautifulsoup4>=4.10.0

# PDF rendering (choose one or both)
PyMuPDF>=1.23.0  # fitz
pdf2image>=1.16.0

# Image processing
Pillow>=9.0.0

# Optional: OCR fallback
pytesseract>=0.3.10

# Streamlit
streamlit>=1.36.0
